requirements.txt
ffastapi==0.75.0
uvicorn[standard]==0.17.6
langdetect==1.0.9
spacy==3.7.2
pydantic
python-dotenv
starlette>=0.27.0,<0.28.0
